backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-aiplatform==1.40.0
httpx==0.26.0
slowapi==0.1.9
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
